feat(Course): 강의자가 아닌 사용자가 수정/삭제를 시도를 방지하는 추가

+++ src/main/java/team7/simple/domain/course/service/CourseService.java
@@ -1,6 +1,7 @@
 package team7.simple.domain.course.service;
 
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import team7.simple.domain.course.dto.CourseDetailResponseDto;
@@ -13,11 +14,13 @@ import team7.simple.domain.course.repository.CourseJpaRepository;
 import team7.simple.domain.enroll.entity.Enroll;
 import team7.simple.domain.enroll.service.EnrollService;
 import team7.simple.domain.user.entity.User;
+import team7.simple.global.error.advice.exception.CAccessDeniedException;
 
 import java.util.List;
 
 
 @RequiredArgsConstructor
+@Slf4j
 @Service
 public class CourseService {
     private final CourseJpaRepository courseJpaRepository;
@@ -40,19 +43,26 @@ public class CourseService {
         return new CourseDetailResponseDto(course, attendeeCount);
     }
 
-
     @Transactional
-    public void deleteCourse(Long courseId) {
+    public Long updateCourse(Long courseId, CourseUpdateParam courseUpdateParam, User user) {
         Course course = getCourseById(courseId);
-        courseJpaRepository.delete(course);
-    }
-
-    @Transactional
-    public Long updateCourse(Long courseId, CourseUpdateParam courseUpdateParam) {
-        Course course = getCourseById(courseId);
+        if (!course.getInstructor().equals(user)){
+            log.info("강의자에게만 수정 권한이 있습니다.");
+            throw new CAccessDeniedException();
+        }
         course.setTitle(courseUpdateParam.getTitle());
         course.setSubtitle(courseUpdateParam.getSubtitle());
         return courseId;
+    }
+
+    @Transactional
+    public void deleteCourse(Long courseId, User user) {
+        Course course = getCourseById(courseId);
+        if (!course.getInstructor().equals(user)){
+            log.info("강의자에게만 삭제 권한이 있습니다.");
+            throw new CAccessDeniedException();
+        }
+        courseJpaRepository.delete(course);
     }
 
     public void register(Long courseId, User user) {
